requirements.txt
fastapi==0.74.1
uvicorn[standard]==0.17.5
SQLAlchemy==1.4.31
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==1.1
requests==2.26
python-multipart==0.0.5
pytest==7.0.1
